!FRB CryEngine converted to new time API
Attempt to list most if not all edits that go a bit beyond 'Convert to CTimeValue'.

- !R Removed GetCurAsyncTimeSec() and GetCurTimeSec() due to obfuscation; Also, confusing vs GetTimer()->GetAsyncCurTime()
- !T m_vStreamPreCachePointDefs changed to vector instead of PODArray, cuz of 'expireTime' in precache points.
- !T m_nLastTimeUsed no longer rounded to nearest second.
- !T m_maxTotalLIfe renamed to m_maxTotalLife XD
- !T Adjusted particle lifetime 'isfinite' check to 'IsValid()', functionality remains the same.
- !R Conversion of particle lifespan into CTimeValue() is Debatable, looks like on spawn it's randomized?
  - Precision in such a processor-heavy system is probably not needed.
  - However, with enough fine-tuning it won't matter and the possiblility for more accurate debugging/interesting effects increases.

- !T EntityEvent now has a 'tVal' value for frame-time instead of a float.
- !R Condensed the mess of a ServerTimer to mostly header.
  - Removed 'minimum' frame time of 0.001f....
  - Linked to max-time step CVar instead of hard-coded value.
- !R GetServerTime() moved to CTimer class
- !R Converted some input to UI time, Debatable.

- !RF Animation time converted to 'nTime' mostly.
  - This is debatable in usage, but otherwise it's EXTREMELY hard to tell animation time vs real time apart!
- !F Special kTime which is basically #-of-frames using animation hz.
- !R GetAverageFrameTime() moved to CTimer class

- !T Boost error throw improved a bit.
- !T Disabled multi-media timer resolution changes when window out of focus.  Can cause errors/issues and isn't unified with system boot settings.
- !R Sleep improved overall, along with multimedia timer resolution handling.
- !RF Condensed and improved the base Timer system. No erroneous comments, lots of clarifications, reduced/eliminated excess conversions etc.
   - TimeValue.h now functions as a proper time-type with appropriate accessors/operators.
   - Removed deprecated GetCurrTime()
- !R Frame-rate limiting sleeps condensed to right before GTimer(system)->UpdateOnFrameStart()
- !R UpdateOnFrameStart() location moved higher since char-manager depends on frame-time.
- !T All GetFrameTime(ITimer::ETIMER_UI) changed to GetFrameTime(true)
  - The way it's implemented, it still returns Simulation time just may ignore pauses. Probably not what everyone else thought it does? xd

- !R Removed physics time-step granularity. mpfloat offers more'n enough precision control.

- !R SAnimTime deprecated in favor of CTimeValue standerdization. Animation-specific stuff moved to SAnimData.

- !T Schematic actions STimerDuration() frame/time syntax changed a bit due to ambiguous conflicts.

- !T Range single-step modified a bit to work for both mpfloat and non-mp types.

+++ Code/CryEngine/CryNetwork/Context/History/MementoHistory.cpp
@@ -134,7 +134,7 @@ bool CMementoHistory::SendCurrentValue(const SSendContext& ctx, CSyncContext* pS
 		CTimeValue curTimeValue = ctx.pView->ContextState()->GetLocalPhysicsTime();
 		if (curTimeValue != timeValueStorage)
 		{
-			NetQuickLog(true, 1, "Physics update sent %fms late", (curTimeValue - timeValueStorage).GetMilliSeconds());
+			NetQuickLog(true, 1, "Physics update sent %fms late", (float)(curTimeValue - timeValueStorage).GetMilliSeconds());
 		}
 	}
 #endif
@@ -180,7 +180,7 @@ bool CMementoHistory::ReadCurrentValue(const SReceiveContext& ctx, bool commit)
 		pTimeValue = &timeValueStorage;
 #if LOG_BUFFER_UPDATES
 		if (CNetCVars::Get().LogBufferUpdates)
-			NetLog("[sync] memento history frame time is %f", timeValueStorage.GetMilliSeconds());
+			NetLog("[sync] memento history frame time is %f", (float)timeValueStorage.GetMilliSeconds());
 #endif
 	}
 
@@ -201,7 +201,7 @@ bool CMementoHistory::ReadCurrentValue(const SReceiveContext& ctx, bool commit)
 #if LOG_BUFFER_UPDATES
 	if (CNetCVars::Get().LogBufferUpdates && pTimeValue)
 	{
-		NetLog("[sync] memento history reads time %f", pTimeValue->GetMilliSeconds());
+		NetLog("[sync] memento history reads time %f", (float)pTimeValue->GetMilliSeconds());
 	}
 #endif
